Refactor shared graph and narrative types
