+++ src/forge/lib/utils/forge-flow-helpers.ts
@@ -12,6 +12,7 @@ import type {
   ForgeGraphKind,
   ForgeReactFlowJson,
 } from '@/forge/types/forge-graph';
+import { createFlowNode } from '@/shared/utils/forge-graph-helpers';
 import { 
   FORGE_CONDITIONAL_BLOCK_TYPE,
   FORGE_EDGE_KIND,
@@ -53,60 +54,7 @@ export function isLinearNodeType(type?: string): boolean {
   return !!type && LINEAR_TYPES.has(type);
 }
 
-export function createFlowNode(
-  type: ForgeNodeType,
-  id: string,
-  x: number,
-  y: number,
-): ForgeReactFlowNode {
-  const base: ForgeNode = {
-    id,
-    type,
-    label: undefined,
-    speaker: undefined,
-    characterId: undefined,
-    content: '',
-    setFlags: undefined,
-    choices: undefined,
-    conditionalBlocks: undefined,
-    storyletCall: undefined,
-    actId: undefined,
-    chapterId: undefined,
-    pageId: undefined,
-    defaultNextNodeId: undefined,
-  };
-
-  if (type === FORGE_NODE_TYPE.CHARACTER) {
-    base.content = 'New dialogue...';
-    base.speaker = 'Character';
-  }
-
-  if (type === FORGE_NODE_TYPE.PLAYER) {
-    const c: ForgeChoice = { id: `c_${Date.now()}`, text: 'Choice 1', nextNodeId: undefined };
-    base.choices = [c];
-  }
-
-  if (type === FORGE_NODE_TYPE.CONDITIONAL) {
-    const b: ForgeConditionalBlock = {
-      id: `block_${Date.now()}`,
-      type: FORGE_CONDITIONAL_BLOCK_TYPE.IF,
-      condition: [],
-      content: '',
-      speaker: undefined,
-      characterId: undefined,
-      nextNodeId: undefined,
-      setFlags: undefined,
-    };
-    base.conditionalBlocks = [b];
-  }
-
-  return {
-    id,
-    type,
-    position: { x, y },
-    data: base,
-  } as ForgeReactFlowNode;
-}
+export { createFlowNode };
 
 export function addChoiceToNodeData(nodeData: ForgeNode): ForgeNode {
   if (nodeData.type !== FORGE_NODE_TYPE.PLAYER) return nodeData;
@@ -406,42 +354,7 @@ export function edgeStrokeColor(edge: ForgeReactFlowEdge, sourceType?: string): 
  * Used for initializing default graphs when none exist.
  * Creates a truly empty graph with no nodes - the first node added will become the start node.
  */
-export function createEmptyForgeGraphDoc(opts: {
-  projectId: number
-  kind: ForgeGraphKind
-  title?: string
-  graphId?: number // Optional ID for generating default title
-}): ForgeGraphDoc {
-  const now = new Date().toISOString()
-  
-  // Generate default title: "New Graph" + first 4 digits of ID (or timestamp if no ID)
-  const generateDefaultTitle = () => {
-    const idStr = opts.graphId ? String(opts.graphId) : String(Date.now())
-    const firstFour = idStr.slice(0, 4)
-    return `New Graph ${firstFour}`
-  }
-  
-  const defaultTitle = opts.title ?? generateDefaultTitle()
-  
-  const emptyFlow: ForgeReactFlowJson = {
-    nodes: [],
-    edges: [],
-    viewport: { x: 0, y: 0, zoom: 1 },
-  }
-  
-  return {
-    id: opts.graphId ?? 0, // Will be assigned by backend when created if 0
-    project: opts.projectId,
-    kind: opts.kind,
-    title: defaultTitle,
-    startNodeId: '', // Empty string for empty graphs - will be set when first node is added
-    endNodeIds: [], // Empty array for empty graphs - will be set when first node is added
-    flow: emptyFlow,
-    compiledYarn: null,
-    updatedAt: now,
-    createdAt: now,
-  }
-}
+export { createEmptyForgeGraphDoc } from '@/shared/utils/forge-graph-helpers';
 
 /**
  * Create a graph with proper start and end nodes.
