+++ src/writer/components/WriterWorkspace/layout/WriterLayout.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import type { ReactNode } from 'react';
-import { FORGE_NODE_TYPE } from '@/forge/types/forge-graph';
+import { FORGE_NODE_TYPE } from '@/shared/types/forge-graph';
 import { useWriterWorkspaceStore } from '@/writer/components/WriterWorkspace/store/writer-workspace-store';
 import { WriterTopBar } from '@/writer/components/WriterWorkspace/layout/WriterTopBar';
